ci: rework checks workflows (#203)
* fix: enable bit operator for BufferUsageFlags

* ci: enable dependabot

* ci: refactor checks

* doc: fix doc alias for init/get_error

* fix: clippy warnings

* fix: clippy errors

* ci: allow not_unsafe_ptr_arg_deref

* fix: failure check for sensor get ID

+++ examples/cursor.rs
@@ -44,7 +44,7 @@ pub fn run(png: &Path) -> Result<(), Box<dyn std::error::Error>> {
                     ..
                 } => break 'mainloop,
                 Event::MouseButtonDown { x, y, .. } => {
-                    canvas.fill_rect(Rect::new(x, y, 1, 1))?;
+                    canvas.fill_rect(Rect::new(x as i32, y as i32, 1, 1))?;
                     canvas.present();
                 }
                 _ => {}
