+++ src/sdl3/gpu/shader.rs
@@ -65,7 +65,7 @@ impl<'a> ShaderBuilder<'a> {
     pub fn with_code(mut self, fmt: ShaderFormat, code: &'a [u8], stage: ShaderStage) -> Self {
         self.inner.format = fmt.0;
         self.inner.code = code.as_ptr();
-        self.inner.code_size = code.len() as usize;
+        self.inner.code_size = code.len();
         self.inner.stage = unsafe { std::mem::transmute(stage as u32) };
         self
     }
